+++ next/src/ui/footer/index.tsx
@@ -4,10 +4,12 @@ export default async function Footer() {
 	const { title } = await getSite()
 
 	return (
-		<footer>
-			<p>
-				&copy; {new Date().getFullYear()} {title}
-			</p>
+		<footer className="border-t p-4 text-center">
+			<div className="mx-auto max-w-screen-xl">
+				<p className="text-sm">
+					&copy; {new Date().getFullYear()} {title}
+				</p>
+			</div>
 		</footer>
 	)
 }