+++ next/src/ui/header/LinkList.tsx
@@ -6,7 +6,7 @@ export default function LinkList({ label, links }: Sanity.LinkList) {
 		<InteractiveDetails className="relative">
 			<summary>{label}</summary>
 
-			<ul className="anim-fade-to-b absolute left-0 top-full min-w-max bg-white">
+			<ul className="anim-fade-to-b absolute left-0 top-full min-w-max border bg-white/90 p-2 backdrop-blur">
 				{links?.map((link, key) => (
 					<li key={key}>
 						<CTA className="link" link={link} />