+++ sanity/schemas/index.ts
@@ -12,10 +12,12 @@ import linkList from './objects/link.list'
 import metadata from './objects/metadata'
 
 // modules
+import blogRollup from './modules/blog-rollup'
+import creativeModule from './modules/creative-module'
+import customHtml from './modules/custom-html'
+import faqList from './modules/faq-list'
 import heroCentered from './modules/hero.centered'
 import heroPostcard from './modules/hero.postcard'
-import blogRollup from './modules/blog-rollup'
-import faqList from './modules/faq-list'
 
 export const schemaTypes = [
 	// documents
@@ -33,6 +35,8 @@ export const schemaTypes = [
 
 	// modules
 	blogRollup,
+	creativeModule,
+	customHtml,
 	faqList,
 	heroCentered,
 	heroPostcard,